[ffx][fastboot] support fractional timeout rates for flashing
Change the timeout rate from u64 to f64 to allow setting rates
lower than 1 MB/s. Also fix a bug where arguments were passed
in the wrong order in flash_partitions.

Test: ffx_flash_tests, ffx_fastboot_lib_test, ffx_fastboot_tool_lib_test
Bug: b/500797346
Fixed: b/500797346
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1582918
Commit-Queue: Colin Nelson <colnnelson@google.com>

+++ src/developer/ffx/lib/fastboot/flash_manifest/src/lib.rs
@@ -33,7 +33,7 @@ pub struct ManifestParams {
     pub no_bootloader_reboot: bool,
     pub skip_verify: bool,
     pub op: Command,
-    pub flash_timeout_rate_mb_per_second: u64,
+    pub flash_timeout_rate_mb_per_second: f64,
     pub flash_min_timeout_seconds: u64,
 }
 
@@ -47,7 +47,7 @@ impl Default for ManifestParams {
             no_bootloader_reboot: false,
             skip_verify: false,
             op: Command::Flash,
-            flash_timeout_rate_mb_per_second: 5000,
+            flash_timeout_rate_mb_per_second: 5000.0,
             flash_min_timeout_seconds: 200,
         }
     }
